requirements.txt
openai-whisper
sounddevice
scipy
pydub
yt-dlp
python-dotenv
google-generativeai